@@ -28,26 +28,26 @@
 		"drizzle-kit": "^0.26.2",
 		"eslint": "^9.13.0",
 		"eslint-config-prettier": "^9.1.0",
-		"eslint-plugin-svelte": "^2.45.1",
+		"eslint-plugin-svelte": "^2.46.0",
 		"globals": "^15.11.0",
 		"markdown-it": "^14.1.0",
 		"prettier": "^3.3.3",
 		"prettier-plugin-svelte": "^3.2.7",
 		"prettier-plugin-tailwindcss": "^0.6.8",
 		"smol-toml": "^1.3.0",
-		"svelte": "^5.0.3",
+		"svelte": "^5.1.2",
 		"svelte-check": "^4.0.5",
 		"tailwindcss": "^3.4.14",
 		"typescript": "^5.6.3",
-		"typescript-eslint": "^8.10.0",
-		"vite": "^5.4.9"
+		"typescript-eslint": "^8.11.0",
+		"vite": "^5.4.10"
 	},
 	"dependencies": {
 		"@lucia-auth/adapter-drizzle": "^1.1.0",
 		"@node-rs/argon2": "^2.0.0",
 		"@sveltejs/adapter-node": "^5.2.8",
-		"better-sqlite3": "^11.4.0",
-		"drizzle-orm": "^0.35.2",
+		"better-sqlite3": "^11.5.0",
+		"drizzle-orm": "^0.35.3",
 		"eventsource-parser": "^3.0.0",
 		"json-schema-to-typescript": "^15.0.2",
 		"katex": "^0.16.11",
